fixed the avatar

+++ controllers/users_controller.js
@@ -48,8 +48,8 @@ module.exports.update=async function(req,res){
                         user.name=req.body.name;
                         user.email=req.body.email;
                         if(req.file){
-                            if(user.avatar){
-                                fs.unlinkSync(path.join(__dirname,'..',user.avatar));
+                            if(user.avatar && fs.existsSync(path.join(__dirname , '..' , user.avatar))){
+                                fs.unlinkSync(path.join(__dirname , '..' , user.avatar));
                             }
                             user.avatar=User.avatarPath+'/'+req.file.filename;
                         }
